+++ src/lib/components/TreeView/TreeItem.tsx
@@ -1,0 +1,71 @@
+import React, { useEffect, useState } from "react";
+import { ChevronDownSolidIcon, ChevronRightSolidIcon } from '../../assets/icons';
+import { Checkbox } from '../Checkbox';
+import { IconButton } from '../IconButton';
+import styles from './TreeItem.module.scss';
+
+export interface ITreeNodeProps {
+	nodeId: string;
+	label: string;
+	children?: any;
+	className?: string;
+	isExpanded?: boolean;
+	isSelected?: boolean;
+	onToggleExpand?: (id: string) => void;
+	onSelect?: (id: string, isSelected: boolean) => void;
+}
+
+export const TreeItem = (props: ITreeNodeProps) => {
+
+	const { nodeId, label, children, className, isExpanded, isSelected, onToggleExpand, onSelect } = props;
+
+	const [_isExpanded, setIsExpanded] = useState<boolean>(false);
+	const [_isSelected, setIsSelected] = useState<boolean>(false);
+
+	const getCssClasses = () => {
+		const cssClasses: Array<string> = [];
+		cssClasses.push(styles.treeItem);
+		className && cssClasses.push(className);
+		return cssClasses.filter(css => css).join(' ');
+	};
+
+	useEffect(() => {
+		if (isExpanded !== undefined)
+			setIsExpanded(isExpanded);
+	}, [isExpanded]);
+
+	useEffect(() => {
+		if (isSelected !== undefined)
+			setIsSelected(isSelected);
+	}, [isSelected]);
+
+	const handleOnToggleExpand = (nodeId: string) => {
+		setIsExpanded(!_isExpanded);
+		onToggleExpand && onToggleExpand(nodeId);
+	}
+
+	const handleOnSelect = (nodeId: string) => {
+		setIsSelected(!_isSelected);
+		onSelect && onSelect(nodeId, !_isSelected);
+	}
+
+	return (
+		<li
+			className={getCssClasses()}
+			style={{ paddingLeft: `${(48 * (children ? 0 : 1))}px` }}
+		>
+			<div className="d-flex align-items-center">
+
+				{children &&
+					<IconButton onClick={() => handleOnToggleExpand(nodeId)} icon={!_isExpanded ? <ChevronRightSolidIcon /> : <ChevronDownSolidIcon />} />
+				}
+
+				<Checkbox checked={_isSelected} onChange={() => handleOnSelect(nodeId)} />
+
+				{label}
+			</div>
+
+			{children && _isExpanded ? <ul>{children}</ul> : null}
+		</li>
+	);
+}